+++ client/src/components/ProductList/ProductList.tsx
@@ -4,8 +4,8 @@ import styled from 'styled-components';
 import Product from '../../models/product';
 import { SortButton } from '../../containers/ProductListContainer';
 import { ProductListOrder } from '../../types/product';
-import { range } from '../../utils/range';
 import ProductListHeader from './ProductListHeader';
+import PageNav from './PageNav';
 import EmptyProducts from './EmptyProducts';
 
 const Container = styled.div`
@@ -24,28 +24,6 @@ const ProductListWrapper = styled.ul`
   gap: 24px;
 
   padding: 0;
-`;
-
-const PageNav = styled.ul`
-  list-style: none;
-
-  display: flex;
-  justify-content: center;
-
-  padding: 0;
-`;
-
-type PageNavItemProps = {
-  isSelected: boolean;
-};
-
-const PageNavItem = styled.li<PageNavItemProps>`
-  padding: 5px;
-  color: ${(props) => (props.isSelected ? props.theme.color.black : props.theme.color.grey4)};
-  cursor: pointer;
-  &:hover {
-    color: ${(props) => props.theme.color.black};
-  }
 `;
 
 type Props = {
@@ -77,20 +55,6 @@ const ProductList = (props: Props): JSX.Element => {
     <ProductItem key={product.id} product={product} onWishClick={getWishClickHandler(product)} />
   ));
 
-  const PageNavItems = range(totalPageCount).map((index) => {
-    const pageNum = index + 1;
-    return (
-      <PageNavItem
-        key={pageNum}
-        isSelected={currentPage === pageNum}
-        onClick={onClickPageNum(pageNum)}
-        data-testid={`pageNav${pageNum}`}
-      >
-        {pageNum}
-      </PageNavItem>
-    );
-  });
-
   return (
     <Container>
       <ProductListHeader
@@ -104,7 +68,11 @@ const ProductList = (props: Props): JSX.Element => {
       ) : (
         <ProductListWrapper>{ProductItems}</ProductListWrapper>
       )}
-      <PageNav>{PageNavItems}</PageNav>
+      <PageNav
+        currentPage={currentPage}
+        totalPageCount={totalPageCount}
+        onClickPageNum={onClickPageNum}
+      />
     </Container>
   );
 };
